--- apps/hyperdrive-trading/src/ui/hyperdrive/hooks/usePreviewCloseLongOld.ts
@@ -1,77 +1,0 @@
-import { HyperdriveABI } from "@hyperdrive/core";
-import { useQuery } from "@tanstack/react-query";
-import { Hyperdrive } from "src/appconfig/types";
-import { QueryStatusType } from "src/ui/base/types";
-import { Address, useAccount, usePublicClient } from "wagmi";
-
-interface UsePreviewCloseLongOptions {
-  market: Hyperdrive;
-  tokenID: bigint | undefined;
-  bondAmountIn: bigint | undefined;
-  minBaseAmountOut: bigint | undefined;
-  destination: Address | undefined;
-  asUnderlying?: boolean;
-  enabled?: boolean;
-}
-
-interface UsePreviewCloseLongResult {
-  previewCloseLongStatus: QueryStatusType;
-  baseAmountOut: bigint | undefined;
-}
-
-/**
- * @deprecated use usePreviewCloseLong
- */
-export function usePreviewCloseLong({
-  market,
-  tokenID,
-  bondAmountIn,
-  minBaseAmountOut,
-  destination,
-  asUnderlying = true,
-  enabled = true,
-}: UsePreviewCloseLongOptions): UsePreviewCloseLongResult {
-  const publicClient = usePublicClient();
-  const { address: account } = useAccount();
-
-  const queryEnabled =
-    !!tokenID &&
-    !!bondAmountIn &&
-    !!minBaseAmountOut &&
-    !!destination &&
-    !!publicClient &&
-    !!account &&
-    enabled;
-
-  const { data, status } = useQuery({
-    queryKey: [
-      "preview-close-long",
-      market.address,
-      bondAmountIn?.toString(),
-      minBaseAmountOut?.toString(),
-      destination?.toString(),
-    ],
-    enabled: queryEnabled,
-    queryFn: queryEnabled
-      ? async () => {
-          const { result } = await publicClient.simulateContract({
-            abi: HyperdriveABI,
-            address: market.address,
-            account,
-            functionName: "closeLong",
-            args: [
-              tokenID,
-              bondAmountIn,
-              minBaseAmountOut,
-              destination,
-              asUnderlying,
-            ],
-          });
-
-          return result;
-        }
-      : undefined,
-  });
-
-  return { baseAmountOut: data, previewCloseLongStatus: status };
-}
