+++ apps/hyperdrive-trading/src/ui/hyperdrive/longs/CloseLongForm/CloseLongForm.tsx
@@ -7,8 +7,8 @@ import { Stat } from "src/ui/base/components/Stat";
 import { formatBalance } from "src/ui/base/formatting/formatBalance";
 import { formatBigInt } from "src/ui/base/formatting/formatBigInt";
 import { useNumericInput } from "src/ui/base/hooks/useNumericInput";
-import { useCloseLong } from "src/ui/hyperdrive/hooks/useCloseLong";
-import { usePreviewCloseLong } from "src/ui/hyperdrive/hooks/usePreviewCloseLong";
+import { useCloseLong } from "src/ui/hyperdrive/longs/hooks/useCloseLong";
+import { usePreviewCloseLong } from "src/ui/hyperdrive/longs/hooks/usePreviewCloseLong";
 import { TokenInput } from "src/ui/token/TokenInput";
 import { parseUnits } from "viem";
 import { useAccount } from "wagmi";
