+++ apps/hyperdrive-trading/src/ui/hyperdrive/longs/CloseLongModalButton/CloseLongModalButton.tsx
@@ -2,7 +2,7 @@ import { XMarkIcon } from "@heroicons/react/24/solid";
 import { Long } from "@hyperdrive/core";
 import { ReactElement } from "react";
 import { Hyperdrive } from "src/appconfig/types";
-import { CloseLongForm } from "src/ui/trade/CloseLongForm/CloseLongForm";
+import { CloseLongForm } from "src/ui/hyperdrive/longs/CloseLongForm/CloseLongForm";
 
 export interface CloseLongModalButtonProps {
   modalId: string;
