+++ packages/hyperdrive/src/amm/events/decodeAssetFromTransferSingleEventData.ts
