+++ packages/hyperdrive/src/amm/longs/getOpenLongs.ts
@@ -2,8 +2,8 @@ import { QueryObserverOptions } from "@tanstack/query-core";
 import groupBy from "lodash.groupby";
 import mapValues from "lodash.mapvalues";
 import { sumBigInt } from "src/base/sumBy";
-import { decodeAssetFromTransferSingleEventData } from "src/amm/assetId";
-import { getTransferSingleEvents } from "src/amm/getTransferSingleEvents";
+import { decodeAssetFromTransferSingleEventData } from "src/amm/events/decodeAssetFromTransferSingleEventData";
+import { getTransferSingleEvents } from "src/amm/events/getTransferSingleEvents";
 import { Long } from "src/amm/longs/types";
 import { PublicClient, Address, Transport, Chain } from "viem";
