+++ packages/hyperdrive/src/amm/shorts/getOpenShorts.ts
@@ -2,8 +2,8 @@ import { QueryObserverOptions } from "@tanstack/query-core";
 import groupBy from "lodash.groupby";
 import mapValues from "lodash.mapvalues";
 import { sumBigInt } from "src/base/sumBy";
-import { decodeAssetFromTransferSingleEventData } from "src/amm/assetId";
-import { getTransferSingleEvents } from "src/amm/getTransferSingleEvents";
+import { decodeAssetFromTransferSingleEventData } from "src/amm/events/decodeAssetFromTransferSingleEventData";
+import { getTransferSingleEvents } from "src/amm/events/getTransferSingleEvents";
 import { PublicClient, Address, Transport, Chain } from "viem";
 import { Short } from "src/amm/shorts/types";
